Adhere to recommendation that clusters begin with video keyframe
Although not a requirement, it is recommended that clusters start
with a video keyframe.  This makes seeking easier and should improve
scrubbing performance in some apps.  It is also known that the
twonky dlna server has a bug and can not handle streams that are not
formatted this way.

I have used a similar technique to mkclean for determining cluster
boundaries.  A new cluster is started at every video keyframe unless
that would place the new cluster closer than 2 seconds to the previous
cluster.  I have also kept the original code that limited cluster
sizes to 20 seconds.  So if we mux streams with very long keyframe
intervals or PIR, the old behavior will kick in.